+++ tests/StatsTest.php
@@ -4,41 +4,13 @@ namespace Cosmastech\LaravelStatsDAdapter\Tests;
 
 use Cosmastech\LaravelStatsDAdapter\AdapterManager;
 use Cosmastech\LaravelStatsDAdapter\Stats;
-use Cosmastech\StatsDClientAdapter\Adapters\Datadog\DatadogStatsDClientAdapter;
 use Cosmastech\StatsDClientAdapter\Adapters\InMemory\InMemoryClientAdapter;
-use Cosmastech\StatsDClientAdapter\Clients\Datadog\DatadogLoggingClient;
 use DateTimeImmutable;
 use Illuminate\Support\Carbon;
-use Illuminate\Support\Facades\Config;
 use PHPUnit\Framework\Attributes\Test;
 
 class StatsTest extends AbstractTestCase
 {
-    #[Test]
-    public function getDefaultInstance_returnsConfigDefaultAdapter()
-    {
-        // Given
-        Config::set("statsd-adapter.default", "this-does-not-exist-but-thats-ok");
-
-        // When
-        $defaultInstanceString = Stats::getDefaultInstance();
-
-        // Then
-        self::assertEquals("this-does-not-exist-but-thats-ok", $defaultInstanceString);
-    }
-
-    #[Test]
-    public function setDefaultInstance_overridesConfigDefault()
-    {
-        // Given application is configured
-
-        // When
-        Stats::setDefaultInstance("yooooo");
-
-        // Then
-        self::assertEquals("yooooo", Stats::getDefaultInstance());
-    }
-
     #[Test]
     public function getFacadeRoot_returnsAdapterManager()
     {
@@ -49,20 +21,6 @@ class StatsTest extends AbstractTestCase
 
         // Then
         self::assertInstanceOf(AdapterManager::class, $facadeRoot);
-    }
-
-    #[Test]
-    public function memoryAdapter_instance_returnsInMemoryClientAdapter()
-    {
-        // Given
-        Config::set("statsd-adapter.default", "memory");
-
-        // When
-        /** @var InMemoryClientAdapter $inMemoryClientAdapter */
-        $inMemoryClientAdapter = Stats::instance();
-
-        // Then
-        self::assertInstanceOf(InMemoryClientAdapter::class, $inMemoryClientAdapter);
     }
 
     #[Test]
@@ -85,18 +43,5 @@ class StatsTest extends AbstractTestCase
             $stats->count[0]->recordedAt->getTimestamp(),
             1
         );
-    }
-
-    #[Test]
-    public function logDatadog_instance_returnsConfiguredDatadogClient()
-    {
-        // Given config set up with log_datadog channel
-
-        // When
-        $datadogClientAdapter = Stats::instance("log_datadog");
-
-        // Then
-        self::assertInstanceOf(DatadogStatsDClientAdapter::class, $datadogClientAdapter);
-        self::assertInstanceOf(DatadogLoggingClient::class, $datadogClientAdapter->getClient());
     }
 }
